+++ how-to/workspace-platform-starter/client/src/modules/actions/window-platform/index.ts
@@ -1,9 +1,9 @@
 import type { ModuleImplementation, ModuleTypes } from "workspace-platform-starter/shapes/module-shapes";
-import { WindowActionsProvider } from "./actions";
+import { WindowPlatformActionsProvider } from "./actions";
 
 /**
  * Define the entry points for the module.
  */
 export const entryPoints: { [type in ModuleTypes]?: ModuleImplementation } = {
-	actions: new WindowActionsProvider()
+	actions: new WindowPlatformActionsProvider()
 };