+++ how-to/workspace-platform-starter/client/src/modules/composite/default-workspace/actions.ts
@@ -1,10 +1,13 @@
+import type {
+	CustomActionPayload,
+	CustomActionsMap,
+	WorkspacePlatformModule
+} from "@openfin/workspace-platform";
 import {
 	CustomActionCallerType,
-	type CustomActionPayload,
-	type CustomActionsMap,
-	type WorkspacePlatformModule
-} from "@openfin/workspace-platform";
-import type { ActionHelpers, Actions } from "workspace-platform-starter/shapes/actions-shapes";
+	type ActionHelpers,
+	type Actions
+} from "workspace-platform-starter/shapes/actions-shapes";
 import type { Logger, LoggerCreator } from "workspace-platform-starter/shapes/logger-shapes";
 import type { ModuleDefinition } from "workspace-platform-starter/shapes/module-shapes";
 import { isEmpty } from "workspace-platform-starter/utils";